Move `MainQueueDispatchDecorator` to it own file
