+++ HackrNewsiOSTests/Live Hackr News/LiveHackrNewsViewControllerTests.swift
@@ -277,11 +277,11 @@ final class LiveHackrNewsViewControllerTests: XCTestCase {
         XCTAssertNil(view?.createdAtLabel.text)
         XCTAssertNil(view?.commentsLabel.text)
     }
-    
+
     func test_loadLiveHackrNewsCompletion_dispatchesFromBackgroundToMainThread() {
         let (sut, loader) = makeSUT()
         sut.loadViewIfNeeded()
-        
+
         let exp = expectation(description: "Wait for loader completion")
         DispatchQueue.global().async {
             loader.completeLiveHackrNewsLoading()
@@ -289,16 +289,16 @@ final class LiveHackrNewsViewControllerTests: XCTestCase {
         }
         wait(for: [exp], timeout: 1.0)
     }
-    
+
     func test_loadLiveHackrNewsCompletion_dispatchesStoryLoaderFromBackgroundToMainThread() {
         let (sut, loader) = makeSUT()
         let lhn0 = makeLiveHackrNew(id: 0).model
         let story0 = makeStory().model
         sut.loadViewIfNeeded()
-        
+
         loader.completeLiveHackrNewsLoading(with: [lhn0], at: 0)
         sut.simulateStoryViewVisible(at: 0)
-        
+
         let exp = expectation(description: "Wait for loader completion")
         DispatchQueue.global().async {
             loader.completeStoryLoading(with: story0, at: 0)
